+++ classes/renderers/renderable/section.php
@@ -59,7 +59,7 @@ class format_ludic_section extends format_ludic_item {
 
         // Action.
         $domain = $contexthelper->get_domain();
-        if (($domain === 'course' || $domain === 'section') && !$contexthelper->is_student_view_forced()) {
+        if (($domain === 'course' || $domain === 'section') && !$contexthelper->is_student_view_forced() && !$contexthelper->is_single_section()) {
             $this->action = 'getDataLinkAndRedirectTo';
             $this->link   = $CFG->wwwroot . '/course/view.php?id=' . $section->courseid . '&section=' . $section->section;
         }