+++ CHANGELOG.md
@@ -1,10 +1,29 @@
 # LUDIC course format : Release Notes
 
-## 3.5.1 - [to be released]
+## 3.5.2 - 2020-11-05
 
 ### ADDED
 
-- Support for Moodle 3.9
+- New ACHIEVEMENT course module skin type added
+- Special student view mode added for single section courses
+- Added rocket-story section skin
+- Added discrete grade-based score activity skin
+
+### CHANGED
+
+- The editor widget for setting weight values for course module instances is now a free text value
+- The PROGRESSION section skin now exposes a 'target' value that can be set on a per-instance basis
+- A few tweeks were made to improve page layout in different places
+
+### FIXED
+
+- Fix behavior of course editing toggle, removing student view button that had side effects
+
+## 3.5.1 - 2020-10-07
+
+### ADDED
+
+- Support for Moodle 3.9 - fixing issue with new resource and activity selection popup
 
 ### CHANGED